+++ scorex-basics/src/main/scala/scorex/network/PeerConnectionHandler.scala
@@ -2,13 +2,14 @@ package scorex.network
 
 import java.net.InetSocketAddress
 
-import akka.actor.{Actor, ActorRef}
+import akka.actor.{SupervisorStrategy, Actor, ActorRef}
 import akka.io.Tcp
 import akka.io.Tcp._
 import akka.util.{ByteString, CompactByteString}
 import com.google.common.primitives.Ints
 import scorex.app.Application
-import scorex.network.NetworkController.PeerHandshake
+import scorex.network.peer.PeerManager
+import scorex.network.peer.PeerManager.Handshaked
 import scorex.utils.ScorexLogging
 
 import scala.util.{Failure, Success}
@@ -21,6 +22,10 @@ case class ConnectedPeer(address: InetSocketAddress, handlerRef: ActorRef) {
   override def equals(obj: scala.Any): Boolean = obj.cast[ConnectedPeer].exists(_.address == this.address)
 }
 
+
+case object Ack extends Event
+
+
 //todo: timeout on Ack waiting
 case class PeerConnectionHandler(application: Application,
                                  connection: ActorRef,
@@ -29,54 +34,63 @@ case class PeerConnectionHandler(application: Application,
 
   import PeerConnectionHandler._
 
-  context watch connection
-
   private lazy val networkControllerRef: ActorRef = application.networkController
+
+  private lazy val peerManager: ActorRef = application.peerManager
 
   val selfPeer = new ConnectedPeer(remote, self)
 
-  private def processErrors: Receive = {
+
+  context watch connection
+
+  override def preStart: Unit = connection ! ResumeReading
+
+  // there is not recovery for broken connections
+  override val supervisorStrategy = SupervisorStrategy.stoppingStrategy
+
+  private def processErrors(stateName: String): Receive = {
     case CommandFailed(w: Write) =>
-      log.info(s"Write failed : $w " + remote)
+      log.warn(s"Write failed :$w " + remote + s" in state $stateName")
       //todo: blacklisting
       //peerManager.blacklistPeer(remote)
-      connection ! Close
+      //connection ! Close
+
+      connection ! ResumeReading
 
     case cc: ConnectionClosed =>
-      networkControllerRef ! NetworkController.PeerDisconnected(remote)
-      log.info("Connection closed to : " + remote + ": " + cc.getErrorCause)
+      peerManager ! PeerManager.Disconnected(remote)
+      log.info("Connection closed to : " + remote + ": " + cc.getErrorCause + s" in state $stateName")
 
     case CloseConnection =>
-      log.info(s"Enforced to abort communication with: " + remote)
+      log.info(s"Enforced to abort communication with: " + remote + s" in state $stateName")
       connection ! Close
+
+    case CommandFailed(cmd: Tcp.Command) =>
+      log.info("Failed to execute command : " + cmd + s" in state $stateName")
+      connection ! ResumeReading
   }
 
-  private def processOwnHandshake(newCycle: Receive): Receive = ({
+  private var handshakeGot = false
+  private var handshakeSent = false
+
+  private object HandshakeCheck
+
+  private def handshake: Receive = ({
     case h: Handshake =>
       connection ! Write(ByteString(h.bytes))
       log.info(s"Handshake sent to $remote")
-      context become newCycle
-  }: Receive) orElse processErrors
+      handshakeSent = true
+      self ! HandshakeCheck
 
-  private def processHandshakeAck(newCycle: Receive): Receive = ({
-    case Received(data) if data.length == HandShakeAck.messageSize =>
-      if (data == HandShakeAck.bytes.toSeq) {
-        log.info(s"Got Handshake Ack from $remote")
-        context become newCycle
-      } else {
-        connection ! Close
-      }
-  }: Receive) orElse processErrors
-
-  private def processHandshake(newCycle: Receive): Receive = ({
-    case Received(data) if data.length > HandShakeAck.messageSize =>
+    case Received(data) =>
       Handshake.parse(data.toArray) match {
         case Success(handshake) =>
           if (handshake.fromNonce != ownNonce) {
-            connection ! Write(HandShakeAck.bytesAsByteString)
-            networkControllerRef ! PeerHandshake(remote, handshake)
+            peerManager ! Handshaked(remote, handshake)
             log.info(s"Got a Handshake from $remote")
-            context become newCycle
+            connection ! ResumeReading
+            handshakeGot = true
+            self ! HandshakeCheck
           } else {
             connection ! Close
           }
@@ -84,7 +98,14 @@ case class PeerConnectionHandler(application: Application,
           log.info(s"Error during parsing a handshake: $t")
           connection ! Close
       }
-  }: Receive) orElse processErrors
+
+    case HandshakeCheck =>
+      if (handshakeGot && handshakeSent) {
+        connection ! ResumeReading
+        context become workingCycle
+      }
+  }: Receive) orElse processErrors(CommunicationState.AwaitingHandshake.toString)
+
 
   def workingCycleLocalInterface: Receive = {
     case msg: message.Message[_] =>
@@ -120,23 +141,28 @@ case class PeerConnectionHandler(application: Application,
             true
         }
       }
-
-    case CommandFailed(cmd: Tcp.Command) =>
-      log.info("Failed to execute command : " + cmd)
-
-    case nonsense: Any => log.warn(s"Strange input for PeerConnectionHandler: $nonsense")
+      connection ! ResumeReading
   }
 
-  def workingCycle: Receive = workingCycleLocalInterface orElse workingCycleRemoteInterface orElse processErrors
+  def workingCycle: Receive =
+    workingCycleLocalInterface orElse
+      workingCycleRemoteInterface orElse
+      processErrors(CommunicationState.WorkingCycle.toString) orElse ({
+      case nonsense: Any =>
+        log.warn(s"Strange input for PeerConnectionHandler: $nonsense")
+    }: Receive)
 
-  override def receive: Receive =
-    processOwnHandshake(
-      processHandshakeAck(processHandshake(workingCycle))
-        orElse processHandshake(processHandshakeAck(workingCycle))
-    ) orElse processHandshake(processOwnHandshake(processHandshakeAck(workingCycle)))
+  override def receive: Receive = handshake
 }
 
 object PeerConnectionHandler {
+
+  private object CommunicationState extends Enumeration {
+    type CommunicationState = Value
+
+    val AwaitingHandshake = Value("AwaitingHandshake")
+    val WorkingCycle = Value("WorkingCycle")
+  }
 
   case object CloseConnection