+++ scorex-transaction/src/main/scala/scorex/transaction/state/LagonakiState.scala
@@ -1,5 +1,0 @@
-package scorex.transaction.state
-
-import scorex.transaction.{AccountTransactionsHistory, BalanceSheet, State}
-
-trait LagonakiState extends State with BalanceSheet with AccountTransactionsHistory