+++ src/main/scala/scorex/lagonaki/api/http/DebugApiRoute.scala
@@ -8,6 +8,7 @@ import play.api.libs.json.Json
 import scorex.api.http._
 import scorex.block.Block.BlockId
 import scorex.crypto.encode.Base58
+import scorex.crypto.hash.FastCryptographicHash
 import scorex.lagonaki.server.LagonakiApplication
 import spray.routing.Route
 
@@ -19,7 +20,28 @@ case class DebugApiRoute(application: LagonakiApplication)(implicit val context:
   lazy val wallet = application.wallet
 
   override lazy val route = pathPrefix("debug") {
-    state ~ stateAt
+    blocks ~ state ~ stateAt
+  }
+
+  @Path("/blocks/{howMany}")
+  @ApiOperation(value = "Blocks", notes = "Get sizes and full hashes for last blocks", httpMethod = "GET")
+  @ApiImplicitParams(Array(
+    new ApiImplicitParam(
+      name = "howMany",
+      value = "How many last blocks to take",
+      required = true,
+      dataType = "String",
+      paramType = "path")
+  ))
+  def blocks: Route = {
+    path("blocks" / IntNumber) { case howMany =>
+      jsonRoute {
+        Json.arr(application.blockStorage.history.lastBlocks(howMany).map { block =>
+          val bytes = block.bytes
+          Json.obj(bytes.length.toString -> Base58.encode(FastCryptographicHash(bytes)))
+        }).toString()
+      }
+    }
   }
 
   @Path("/state")
@@ -44,7 +66,7 @@ case class DebugApiRoute(application: LagonakiApplication)(implicit val context:
     path("state" / Segment) { case blockId =>
       jsonRoute {
         val id: BlockId = Base58.decode(blockId).getOrElse(Array.empty)
-        application.blockStorage.state(Some(id)) match {
+        application.blockStorage.state(id) match {
           case None => Json.obj("error" -> "wrong block id").toString
           case Some(b) => b.toString
         }