+++ src/main/scala/scorex/lagonaki/api/http/PeersHttpService.scala
@@ -1,5 +1,6 @@
 package scorex.lagonaki.api.http
 
+import java.net.InetSocketAddress
 import javax.ws.rs.Path
 
 import akka.actor.ActorRefFactory
@@ -8,7 +9,8 @@ import com.wordnik.swagger.annotations._
 import play.api.libs.json.{JsString, Json}
 import scorex.api.http.{ApiRoute, CommonApiFunctions}
 import scorex.lagonaki.server.LagonakiApplication
-import scorex.network.{Handshake, NetworkController}
+import scorex.network.Handshake
+import scorex.network.peer.PeerManager
 import spray.http.MediaTypes._
 import spray.routing.Route
 
@@ -21,25 +23,52 @@ case class PeersHttpService(application: LagonakiApplication)(implicit val conte
 
   override lazy val route =
     pathPrefix("peers") {
-      peers // TODO implement and fix ~ score
+      allPeers ~ connectedPeers // TODO implement and fix ~ score
     }
 
-  @Path("/")
+  @Path("/all")
   @ApiOperation(value = "Peer list", notes = "Peer list", httpMethod = "GET")
   @ApiResponses(Array(
     new ApiResponse(code = 200, message = "Json with peer list or error")
   ))
-  def peers: Route = path("") {
+  def allPeers: Route = path("all") {
     get {
       respondWithMediaType(`application/json`) {
         onComplete {
-          (application.networkController ? NetworkController.GetConnectedPeers).map { handshakes =>
-            val peerData = Json.arr(handshakes.asInstanceOf[Seq[Handshake]].map { handshake =>
-              val s = handshake.fromAddress.toString + ":" + handshake.fromPort + "::" + handshake.fromNonce
-              JsString(s)
-            })
-            Json.obj("peers" -> peerData).toString()
-          }
+          (application.peerManager ? PeerManager.GetAllPeers)
+            .mapTo[Seq[InetSocketAddress]]
+            .map { addresses =>
+              val peerData = Json.arr(addresses.map { address =>
+                JsString(address.toString)
+              })
+              Json.obj("peers" -> peerData).toString()
+            }
+        } {
+          case Success(value) => complete(value)
+          case Failure(ex) => failWith(ex)
+        }
+      }
+    }
+  }
+
+  @Path("/connected")
+  @ApiOperation(value = "Connected peers list", notes = "Connected peers list", httpMethod = "GET")
+  @ApiResponses(Array(
+    new ApiResponse(code = 200, message = "Json with connected peers or error")
+  ))
+  def connectedPeers: Route = path("connected") {
+    get {
+      respondWithMediaType(`application/json`) {
+        onComplete {
+          (application.peerManager ? PeerManager.GetConnectedPeers)
+            .mapTo[Seq[Handshake]]
+            .map { handshakes =>
+              val peerData = Json.arr(handshakes.map { handshake =>
+                val s = handshake.fromAddress.toString + ":" + handshake.fromPort + "::" + handshake.fromNonce
+                JsString(s)
+              })
+              Json.obj("peers" -> peerData).toString()
+            }
         } {
           case Success(value) => complete(value)
           case Failure(ex) => failWith(ex)